pyproject.toml
[tool.poetry]
name = "python-local-env-poetry"
version = "0.1.0"
description = "Python local env with Poetry"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.2"

[tool.poetry.dev-dependencies]
google-cloud-bigquery = "3.27.0"
dacite = "1.6.0"
toolz = "0.12.0"
ndjson = "0.3.1"
pytest = "8.3.4"
